Numba for q_hat_matrix
